requirements.txt
asgiref==3.3.1
cycler==0.10.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.5
django-admin-interface==0.15.5
django-colorfield==0.4.1
django-cors-headers==3.7.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-import-export==2.5.0
djangorestframework==3.12.2
et-xmlfile==1.0.1
gunicorn==20.0.4
kiwisolver==1.3.1
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.1.0
psycopg2-binary==2.8.6
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
PyYAML==5.4.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
